Moved out AutonoteData
