+++ src/Engine/Data/AutonoteData.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+#include "AutonoteEnums.h"
+
+struct AutonoteData {
+    std::string pText;
+    AutonoteType eType;
+};
